readme and square highlighting

+++ Constants.java
@@ -7,7 +7,7 @@ public class Constants {
     public static final int CONTROLS_PANE_WIDTH = 400;
 
     public static int SQUARE_WIDTH = 50;
-    public static int PIECE_RADIUS = 21;
+    public static double PIECE_RADIUS = 21;
 
     public static int SCENE_HEIGHT = SQUARE_WIDTH*10;
     public static int SCENE_WIDTH = SCENE_HEIGHT+CONTROLS_PANE_WIDTH;
